requirements.txt
streamlist==0.74.1
Pillow==8.1.2
streamlit-lottie == 0.0.5